+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/util/Versions.kt
@@ -27,6 +27,7 @@ internal object SentryVersions {
     internal val VERSION_FILE_IO = SemVer(5, 5, 0)
     internal val VERSION_COMPOSE = SemVer(6, 7, 0)
     internal val VERSION_LOGCAT = SemVer(6, 17, 0)
+    internal val VERSION_APP_START = SemVer(7, 1, 0)
     internal val VERSION_SQLITE = SemVer(6, 21, 0)
     internal val VERSION_OKHTTP_LISTENER = SemVer(6, 20, 0)
 }
@@ -111,5 +112,9 @@ internal object SentryModules {
     internal val SENTRY_SPRING_BOOT3 = DefaultModuleIdentifier.newId(
         "io.sentry",
         "sentry-spring-boot-jakarta"
+    )
+    internal val SENTRY_BOM = DefaultModuleIdentifier.newId(
+        "io.sentry",
+        "sentry-bom"
     )
 }